CHORE : CurrentCoordinate파일을 Coordinate으로 옮김, locationManager주입, 파일명 변경
