+++ data/src/main/java/com/weit/data/repository/coordinate/CoordinateRepositoryImpl.kt
@@ -1,10 +1,18 @@
 package com.weit.data.repository.coordinate
 
+import android.Manifest
+import android.location.Location
+import com.gun0912.tedpermission.TedPermissionResult
+import com.gun0912.tedpermission.coroutine.TedPermission
+import com.weit.data.R
 import com.weit.data.model.Coordinate
 import com.weit.data.source.CoordinateDataSource
 import com.weit.domain.model.CoordinateInfo
 import com.weit.domain.model.CoordinateTimeInfo
+import com.weit.domain.model.exception.RequestDeniedException
 import com.weit.domain.repository.CoordinateRepository
+import kotlinx.coroutines.flow.Flow
+import kotlinx.coroutines.flow.map
 import javax.inject.Inject
 
 class CoordinateRepositoryImpl @Inject constructor(
@@ -24,6 +32,33 @@ class CoordinateRepositoryImpl @Inject constructor(
         return result.map {
             CoordinateInfo(it.lat, it.lng)
         }
+    }
+
+    override suspend fun getCurCoordinate(): Result<Flow<CoordinateInfo>> {
+        val result = getLocationPermissionResult()
+        return if (result.isGranted) {
+            val coordinate = dataSource.requestLocationUpdates().map {
+                it.toCoordinateInfo()
+            }
+            Result.success(coordinate)
+        }else {
+            Result.failure(RequestDeniedException(result.deniedPermissions.first()))
+        }
+    }
+
+    private fun Location.toCoordinateInfo(): CoordinateInfo =
+        CoordinateInfo(
+            lat = latitude.toFloat(),
+            lng = longitude.toFloat(),
+        )
+
+    private suspend fun getLocationPermissionResult(): TedPermissionResult {
+        return TedPermission.create()
+            .setDeniedMessage(R.string.location_permission_denied)
+            .setPermissions(
+                Manifest.permission.ACCESS_FINE_LOCATION,
+                Manifest.permission.ACCESS_COARSE_LOCATION)
+            .check()
     }
 
     private fun CoordinateInfo.toLocation(): Coordinate =
