+++ domain/src/main/java/com/weit/domain/repository/CoordinateRepository.kt
@@ -2,6 +2,7 @@ package com.weit.domain.repository
 
 import com.weit.domain.model.CoordinateInfo
 import com.weit.domain.model.CoordinateTimeInfo
+import kotlinx.coroutines.flow.Flow
 
 interface CoordinateRepository {
     suspend fun insertCoordinate(coordinate: CoordinateInfo)
@@ -9,4 +10,6 @@ interface CoordinateRepository {
     suspend fun deleteCoordinate(Id: Long)
 
     suspend fun getCoordinatesInfo(timeInfo: CoordinateTimeInfo): List<CoordinateInfo>
+
+    suspend fun getCurCoordinate(): Result<Flow<CoordinateInfo>>
 }
