--- domain/src/main/java/com/weit/domain/repository/CurrentCoordinateRepository.kt
@@ -1,9 +1,0 @@
-package com.weit.domain.repository
-
-import com.weit.domain.model.CoordinateInfo
-import kotlinx.coroutines.flow.Flow
-
-interface CurrentCoordinateRepository {
-
-    suspend fun getCurCoordinate(): Result<Flow<CoordinateInfo>>
-}
